requirements.txt
streamlit
firebase-admin
requests
pandas
reportlab
arabic-reshaper
python-bidi
